refactor: relocate placeholder users feature
