--- lib/domain/facades/users_repo.dart
@@ -1,7 +1,0 @@
-import '../use_cases/get_users/use_case.dart';
-
-abstract class UsersRepo {
-  const UsersRepo();
-
-  Future<FailureOrUsers> getUsers();
-}
