--- lib/domain/use_cases/get_users/failures.dart
@@ -1,6 +1,0 @@
-part of 'use_case.dart';
-
-@freezed
-class GetUsersFailure with _$GetUsersFailure {
-  const factory GetUsersFailure.offline() = _GetUsersFailureOffline;
-}
