--- lib/domain/use_cases/types.dart
@@ -1,7 +1,0 @@
-import 'package:dartz/dartz.dart';
-
-abstract class FutureUseCase<FailureType, ReturnType, ParamsType> {
-  const FutureUseCase();
-
-  Future<Either<FailureType, ReturnType>> call(ParamsType params);
-}
