+++ lib/features/users_placeholder/application/users_cubit/state.dart
@@ -1,0 +1,20 @@
+part of 'cubit.dart';
+
+@freezed
+class UsersFailure with _$UsersFailure {
+  const factory UsersFailure.onGet(
+    GetUsersFailure failure,
+  ) = _UsersFailureOnGet;
+}
+
+@freezed
+class UsersState with _$UsersState {
+  const factory UsersState.idle() = _UsersStateIdle;
+  const factory UsersState.loading() = _UsersStateLoading;
+  const factory UsersState.loaded(
+    KtList<User> users,
+  ) = _UsersStateLoaded;
+  const factory UsersState.failed(
+    UsersFailure failure,
+  ) = _UsersStateFailed;
+}
