+++ lib/features/users_placeholder/domain/get_users_failure.dart
@@ -1,0 +1,8 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'get_users_failure.freezed.dart';
+
+@freezed
+class GetUsersFailure with _$GetUsersFailure {
+  const factory GetUsersFailure.offline() = _GetUsersFailureOffline;
+}
