+++ lib/features/users_placeholder/domain/user.dart
