+++ lib/features/users_placeholder/infrastructure/converters/user.dart
@@ -1,0 +1,10 @@
+import '../../domain/user.dart';
+import '../dto/user.dart';
+
+extension ConvertibleUserDto on UserDto {
+  User get asEntity => User(
+        id: id,
+        name: name,
+        username: username,
+      );
+}
