+++ lib/features/users_placeholder/infrastructure/dto/user.dart
@@ -1,0 +1,16 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'user.freezed.dart';
+part 'user.g.dart';
+
+@freezed
+class UserDto with _$UserDto {
+  const factory UserDto({
+    required int id,
+    required String name,
+    required String username,
+  }) = _UserDto;
+
+  factory UserDto.fromJson(Map<String, dynamic> json) =>
+      _$UserDtoFromJson(json);
+}
