--- lib/infrastructure/models/user_converters.dart
@@ -1,9 +1,0 @@
-part of 'user.dart';
-
-extension ConvertibleUserR on UserR {
-  User get asEntity => User(
-        id: id,
-        name: name,
-        username: username,
-      );
-}
